Baseline implementation The baseline offers a command line interface to Winston's on-disk music service. The service indexes the provided music library (a tree of directories of format library/artist/album/songs) and an interface to the pygame playback driver.
The main feature is that when a user requests a song via the "play"
command, it need not be an exact path or title match. Instead, the
service scans the available songs and selects the one with the shortest
levenshtein distance to the requested title (levenshtein implementation
from the cer module of HW2).